+++ src/main/java/servlet/pahanaedu/book/dto/BookDTO.java
@@ -1,4 +1,4 @@
-package servlet.pahanaedu.dto;
+package servlet.pahanaedu.book.dto;
 
 public class BookDTO {
     private int id;
