+++ src/main/java/servlet/pahanaedu/category/dto/CategoryDTO.java
@@ -1,4 +1,4 @@
-package servlet.pahanaedu.dto;
+package servlet.pahanaedu.category.dto;
 
 public class CategoryDTO {
     private int id;
