+++ src/main/java/servlet/pahanaedu/servlet/LoginServlet.java
@@ -1,7 +1,7 @@
 package servlet.pahanaedu.servlet;
 
-import servlet.pahanaedu.dto.UserDTO;
-import servlet.pahanaedu.service.UserService;
+import servlet.pahanaedu.user.dto.UserDTO;
+import servlet.pahanaedu.user.service.UserService;
 
 import javax.servlet.ServletException;
 import javax.servlet.annotation.WebServlet;
