--- src/main/java/servlet/pahanaedu/user/m.java
@@ -1,4 +1,0 @@
-package servlet.pahanaedu.user;
-
-public class m {
-}
